+++ workspaces/blackduck/plugins/blackduck/src/components/BlackDuckCard/index.ts
@@ -1,0 +1,1 @@
+export { RiskCardComponent } from './RiskCard';