+++ workspaces/redhat-resource-optimization/plugins/redhat-resource-optimization-backend/src/index.ts
@@ -1,0 +1,2 @@
+export * from './service/router';
+export { resourceOptimizationPlugin as default } from './plugin';