+++ workspaces/redhat-resource-optimization/plugins/redhat-resource-optimization-common/scripts/lib/yaml.mjs
@@ -1,0 +1,6 @@
+import { Document as YamlDocument } from 'yaml';
+
+export function toYaml(jsonObject) {
+  const yaml = new YamlDocument(jsonObject);
+  return yaml.toString();
+}