+++ workspaces/redhat-resource-optimization/plugins/redhat-resource-optimization/src/components/resource-optimization-icon/index.ts
@@ -1,0 +1,3 @@
+export { ResourceOptimizationIcon } from './ResourceOptimizationIcon';
+export { ResourceOptimizationIconOutlined } from './ResourceOptimizationIconOutlined';
+export { ResourceOptimizationIconFilled } from './ResourceOptimizationIconFilled';