+++ workspaces/report-portal/plugins/report-portal/dev/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { createDevApp } from '@backstage/dev-utils';
+import { reportPortalPlugin, ReportPortalGlobalPage } from '../src/plugin';
+
+createDevApp()
+  .registerPlugin(reportPortalPlugin)
+  .addPage({
+    element: <ReportPortalGlobalPage />,
+    title: 'Root Page',
+    path: '/report-portal',
+  })
+  .render();