+++ workspaces/report-portal/plugins/report-portal/src/index.ts
@@ -1,0 +1,8 @@
+export {
+  reportPortalPlugin,
+  ReportPortalOverviewCard,
+  ReportPortalGlobalPage,
+} from './plugin';
+
+export { isReportPortalAvailable } from './utils/isReportPortalAvailable';
+export { ReportPortalIcon } from './components/ReportPortalIcon';